requirements.txt
aiohttp>=3.9.3
asyncpg>=0.29.0
redis>=5.0.3
SQLAlchemy>=2.0.27
pydantic>=2.6.4
pydantic-settings>=2.2.1
python-dotenv>=1.0.1
tenacity>=8.2.3
pandas>=2.2.0
numpy>=1.26.4
scikit-learn>=1.4.1
scipy>=1.11.4
statsmodels>=0.14.1
python-telegram-bot>=21.0
apscheduler>=3.10.4
psycopg2-binary>=2.9.9
fastapi>=0.110.0
uvicorn>=0.29.0
openai>=1.14.0
lightgbm>=4.3.0
playwright>=1.40.0
